Animal muscle work (amw) analysis data constants

A string list containing a selection of amw analysis constants, used in the "amw_analysis_data.R" excel file.

Usage

amw_analysis_constants

Format

A string list with 35 entries.

prop_working_animals_col

The metadata column "Prop.Working.animals" representing the proportion of live animals that are working animals.

prop_wkg_anmls_ag_col

The metadata column "Prop.Working.animals.Ag" representing the proportion of working animals that work in agriculture.

prop_wkg_anmls_tr_col

The metadata column "Prop.Working.animals.Tr" representing the proportion of working animals that work transporting goods outside of agriculture.

wa_perc_sheet

The excel sheet name "WA_perc" representing the sheet containing information on the proportion of working animals.

wa_enduse_sheet

The excel sheet name "WA_enduse" representing the sheet containing information of the proportion of working animals in agriculture/transport.

wa_feed_sheet

The excel sheet name "WA_feed" representing the sheet containing information on the feed requirements of working animals.

wa_days_hours_sheet

The excel sheet name "WA_days_hours" representing the sheet containing information on the number of days and hours worked by working animals.

wa_power_sheet

The excel sheet name "WA_power" representing the sheet containing information on the power output of working animals.

working_seconds_col

The data column "Working.seconds [seconds per animal]" representing the number of seconds of work performed by working animals.

working_hours_col

The data column "Working.hours [hour per animal]" representing the number of hours of work performed by working animals.

working_days_col

The data column "Working.days [day per animal]" representing the number of days that working animals worked.

nonworking_days_col

The data column "Non-Working days [day per animal]" representing the number of days that working animals did not work.

power_per_animal

The data column "Power.per.animal [W]" representing the power output of working animals.

live_animals_col

The data column "Live.animals" representing the number of live animals.

working_animals_col

The data column "Working.animals" representing the number of working animals.

working_animals_total_col

The data column "Working.animals.total" representing the total number of working animals.

working_animals_ag_col

The data column "Working.animals.Ag" representing the number of working animals that work in agriculture.

working_animals_tr_col

The data column "Working.animals.Tr" representing the number of working animals that work transporting goods outside of agriculture.

working_day_feed_col

The data column "Working.day.feed [MJ/day per animal]" representing the quantity of feed required by working animals each working day.

nonworking_day_feed_col

The data column "Non-Working.day.feed [MJ/day per animal]" representing the quantity of feed required by working animals each non-working day.

working_yearly_feed_col

The data column "Working.yearly.feed [MJ/year per animal]" representing the quantity of feed required by working animals on working days each year.

nonwkg_yearly_feed_col

The data column "Non-Working.yearly.feed [MJ/year per animal]" representing the quantity of feed required by working animals on non-working days each year.

total_yearly_feed_col

The data column "Total.yearly.feed [MJ/year per animal]" representing the quantity of feed required by working animals each year.

useful_energy_total

The data column "Useful.energy.total [MJ/year]" representing the total useful energy produced by working animals through muscle work.

useful_energy_ag

The data column "Useful.energy.Ag [MJ/year]" representing the useful energy produced by working animals through muscle work in agriculture.

useful_energy_tr

The data column "Useful.energy.Tr [MJ/year]" representing the useful energy produced by working animals through muscle work transporting goods outside of agriculture.

final_energy_total

The data column "Final.energy.total [MJ/year]" representing the total final energy (feed) required by working animals.

final_energy_ag

The data column "Final.energy.Ag [MJ/year]" representing the final energy (feed) required by working animals in agriculture.

final_energy_tr

The data column "Final.energy.Tr [MJ/year]" representing the final energy (feed) required by working animals working to transport goods outside of agriculture.

primary_energy_total

The data column "Primary.energy.total [MJ/year]" representing the primary energy (total feed crop biomass) embodied in the total final energy required by working animals.

primary_energy_ag

The data column "Primary.energy.Ag [MJ/year]" representing the primary energy (total feed crop biomass) embodied in the final energy required by working animals in agriculture.

primary_energy_tr

The data column "Primary.energy.Tr [MJ/year]" representing the primary energy (total feed crop biomass) embodied in the final energy required by working animals transporting goods outside of agriculture.

amw_region_col

The data column "AMW.Region" representing the animal muscle work region.

metric_col

The metadata column "Metric" representing the metric name associated with a particular data series.

yes_const

The string "Yes" representing when a quantity is present or applicable.

Examples

amw_analysis_constants
#> $prop_working_animals_col
#> [1] "Prop.Working.animals"
#> 
#> $prop_wkg_anmls_ag_col
#> [1] "Prop.Working.animals.Ag"
#> 
#> $prop_wkg_anmls_tr_col
#> [1] "Prop.Working.animals.Tr"
#> 
#> $wa_perc_sheet
#> [1] "WA_perc"
#> 
#> $wa_enduse_sheet
#> [1] "WA_enduse"
#> 
#> $wa_feed_sheet
#> [1] "WA_feed"
#> 
#> $wa_days_hours_sheet
#> [1] "WA_days_hours"
#> 
#> $wa_power_sheet
#> [1] "WA_power"
#> 
#> $working_seconds_col
#> [1] "Working.seconds [seconds per animal]"
#> 
#> $working_hours_col
#> [1] "Working.hours [hour per animal]"
#> 
#> $working_days_col
#> [1] "Working.days [day per animal]"
#> 
#> $nonworking_days_col
#> [1] "Non-Working days [day per animal]"
#> 
#> $power_per_animal
#> [1] "Power.per.animal [W]"
#> 
#> $live_animals_col
#> [1] "Live.animals"
#> 
#> $working_animals_col
#> [1] "Working.animals"
#> 
#> $working_animals_total_col
#> [1] "Working.animals.total"
#> 
#> $working_animals_ag_col
#> [1] "Working.animals.Ag"
#> 
#> $working_animals_tr_col
#> [1] "Working.animals.Tr"
#> 
#> $working_day_feed_col
#> [1] "Working.day.feed [MJ/day per animal]"
#> 
#> $nonworking_day_feed_col
#> [1] "Non-Working.day.feed [MJ/day per animal]"
#> 
#> $working_yearly_feed_col
#> [1] "Working.yearly.feed [MJ/year per animal]"
#> 
#> $nonwkg_yearly_feed_col
#> [1] "Non-Working.yearly.feed [MJ/year per animal]"
#> 
#> $total_yearly_feed_col
#> [1] "Total.yearly.feed [MJ/year per animal]"
#> 
#> $useful_energy_total
#> [1] "Useful.energy.total [MJ/year]"
#> 
#> $useful_energy_ag
#> [1] "Useful.energy.Ag [MJ/year]"
#> 
#> $useful_energy_tr
#> [1] "Useful.energy.Tr [MJ/year]"
#> 
#> $final_energy_total
#> [1] "Final.energy.total [MJ/year]"
#> 
#> $final_energy_ag
#> [1] "Final.energy.Ag [MJ/year]"
#> 
#> $final_energy_tr
#> [1] "Final.energy.Tr [MJ/year]"
#> 
#> $primary_energy_total
#> [1] "Primary.energy.total [MJ/year]"
#> 
#> $primary_energy_ag
#> [1] "Primary.energy.Ag [MJ/year]"
#> 
#> $primary_energy_tr
#> [1] "Primary.energy.Tr [MJ/year]"
#> 
#> $amw_region_col
#> [1] "AMW.Region"
#> 
#> $metric_col
#> [1] "Metric"
#> 
#> $yes_const
#> [1] "Yes"
#>