Calculate the useful energy produced by working animals performing muscle work

Calculate the useful energy produced by working animals performing muscle work

calc_useful_energy(
  .df,
  amw_analysis_data_path = MWTools::amw_analysis_data_path(),
  species = MWTools::mw_constants$species,
  method_source = MWTools::mw_constants$method_source,
  wa_power_sheet = MWTools::amw_analysis_constants$wa_power_sheet,
  wa_days_hours_sheet = MWTools::amw_analysis_constants$wa_days_hours_sheet,
  working_days_col = MWTools::amw_analysis_constants$working_days_col,
  working_hours_col = MWTools::amw_analysis_constants$working_hours_col,
  working_seconds_col = MWTools::amw_analysis_constants$working_seconds_col,
  amw_region_code_col = MWTools::conc_cols$amw_region_code_col,
  power_per_animal = MWTools::amw_analysis_constants$power_per_animal,
  useful_energy_total = MWTools::amw_analysis_constants$useful_energy_total,
  useful_energy_ag = MWTools::amw_analysis_constants$useful_energy_ag,
  useful_energy_tr = MWTools::amw_analysis_constants$useful_energy_tr,
  working_animals_total_col = MWTools::amw_analysis_constants$working_animals_total_col,
  working_animals_ag_col = MWTools::amw_analysis_constants$working_animals_ag_col,
  working_animals_tr_col = MWTools::amw_analysis_constants$working_animals_tr_col
)

Arguments

.df

A data frame containing the primary and final energy consumed by working animals. Usually produced by calling the tidy_fao_live_animals, add_concordance_codes, trim_fao_data, get_working_species, calc_working_animals, calc_sector_split, calc_yearly_feed, calc_final_energy, and calc_primary_energy functions in sequence on the raw FAO data.

amw_analysis_data_path

The path to the animal muscle work analysis data, containing data for the number of working hours and power outputs of working animals. Set to the function MWTools::amw_analysis_data_path() by default, which returns the path the analysis data bundled with the MWTools package.

species, method_source

See MWTools::mw_constants.

wa_power_sheet, wa_days_hours_sheet

See MWTools::amw_analysis_constants.

working_hours_col, working_seconds_col

See MWTools::amw_analysis_constants.

amw_region_code_col, power_per_animal

See MWTools::amw_analysis_constants.

useful_energy_total, useful_energy_ag, useful_energy_tr

See MWTools::amw_analysis_constants.

working_animals_total_col, working_animals_ag_col, working_animals_tr_col, working_days_col

See MWTools::amw_analysis_constants.

Examples

MWTools::amw_test_data_path() %>%
  read.csv() %>%
  tidy_fao_live_animals() %>%
  add_concordance_codes() %>%
  trim_fao_data() %>%
  get_working_species() %>%
  calc_working_animals() %>%
  calc_sector_split() %>%
  calc_yearly_feed() %>%
  calc_final_energy() %>%
  calc_primary_energy() %>%
  calc_useful_energy()
#> # A tibble: 354 × 24
#>    Country.code AMW.Region.code Country.name     Year Species   Live.animals
#>    <chr>        <chr>           <chr>           <dbl> <chr>            <dbl>
#>  1 CHN          CHN             China, mainland  1961 Asses         1000000 
#>  2 CHN          CHN             China, mainland  1961 Buffaloes     1000000 
#>  3 CHN          CHN             China, mainland  1961 Camelids      1000000 
#>  4 CHN          CHN             China, mainland  1961 Cattle        1000000 
#>  5 CHN          CHN             China, mainland  1961 Horses        1000000 
#>  6 CHN          CHN             China, mainland  1961 Mules         1000000 
#>  7 CHN          CHN             China, mainland  1962 Asses          998276.
#>  8 CHN          CHN             China, mainland  1962 Buffaloes      998276.
#>  9 CHN          CHN             China, mainland  1962 Camelids       998276.
#> 10 CHN          CHN             China, mainland  1962 Cattle         998276.
#> # … with 344 more rows, and 18 more variables: Prop.Working.animals <dbl>,
#> #   Working.animals.total <dbl>, Prop.Working.animals.Ag <dbl>,
#> #   Prop.Working.animals.Tr <dbl>, Working.animals.Ag <dbl>,
#> #   Working.animals.Tr <dbl>, `Total.yearly.feed [MJ/year per animal]` <dbl>,
#> #   `Final.energy.total [MJ/year]` <dbl>, `Final.energy.Ag [MJ/year]` <dbl>,
#> #   `Final.energy.Tr [MJ/year]` <dbl>, `Primary.energy.total [MJ/year]` <dbl>,
#> #   `Primary.energy.Ag [MJ/year]` <dbl>, `Primary.energy.Tr [MJ/year]` <dbl>, …