Country concordance columns

A string list containing the names of the columns containing country code information.

Usage

conc_cols

Format

A string list with 9 entries.

country_code_col

The metadata column name "Country.code" containing three letter ISO 3166-1 country codes.

country_col

The metadata column name "Country" containing three letter ISO 3166-1 country codes.

country_incl_col

The metadata column name "Country.incl." containing either "Yes" or "No" depending on whether a particular region or country should be included for analysis in MWTools.

exemplar_country_col

The metadata column name "Exemplar.country", representing the country from which some data should be taken if absent from a particular country.

amw_region_code_col

The metadata column "AMW.Region.code" telling the region for this row.

hmw_region_code_col

The metadata column "HMW.Region.code" telling the region for this row.

region_code_col

The metadata column "Region.code" telling the region for this row.

agg_code_col

The metadata column "Agg.Code", representing the country, or country group code for individual country level data to be aggregated in to.

mapping_sheet

The name of the sheet for name mappings.

species

The name of the species column in the concordance spreadsheet.

Examples

conc_cols
#> $country_code_col
#> [1] "Country.code"
#> 
#> $country_col
#> [1] "Country"
#> 
#> $country_incl_col
#> [1] "Country.incl."
#> 
#> $exemplar_country_col
#> [1] "Exemplar.country"
#> 
#> $amw_region_code_col
#> [1] "AMW.Region.code"
#> 
#> $hmw_region_code_col
#> [1] "HMW.Region.code"
#> 
#> $region_code_col
#> [1] "Region.code"
#> 
#> $mapping_sheet
#> [1] "Mapping"
#> 
#> $species
#> [1] "Species"
#>