Retrieve data with sectors organised by "Broad sector"

Retrieve data with sectors organised by "Broad sector"

Usage

get_broad.sector_data(
  .df,
  sex_ilo_col = MWTools::ilo_cols$sex_ilo_col,
  sector_col = MWTools::mw_constants$sector_col,
  ilo_agr_name = MWTools::hmw_sector_constants$ilo_agr_name,
  ilo_ind_name = MWTools::hmw_sector_constants$ilo_ind_name,
  ilo_ser_name = MWTools::hmw_sector_constants$ilo_ser_name,
  agr_name = MWTools::hmw_sector_constants$agr_name,
  ind_name = MWTools::hmw_sector_constants$ind_name,
  ser_name = MWTools::hmw_sector_constants$ser_name,
  ilo_female_name = MWTools::hmw_sex_constants$ilo_female_name,
  ilo_male_name = MWTools::hmw_sex_constants$ilo_male_name,
  female_name = MWTools::hmw_sex_constants$female_name,
  male_name = MWTools::hmw_sex_constants$male_name
)

Arguments

.df

A data frame containing the number of hours worked. Usually produced by calling the add_hmw_region_codes, fill_ilo_data, and calc_total_hours_worked functions in sequence on the raw FAO data.

sex_ilo_col

See MWTools::ilo_cols.

sector_col

See mWTools::mw_constants.

ilo_agr_name, ilo_ind_name, ilo_ser_name, agr_name, ind_name, ser_name

String names in .df and the outgoing data frame for the ILO economic sectors of interest to MW calculations.

ilo_female_name, ilo_male_name, female_name, male_name

String names in .df and the outgoing data frame for the ILO sex identifiers of interest to MW calculations. For now, we include only males and females. There are instances of "other" that will need to be dealt with in future releases.

Examples

ilo_working_hours_data <- read.csv(file = MWTools::ilo_working_hours_test_data_path())
ilo_employment_data <- read.csv(file = MWTools::ilo_employment_test_data_path())
hmw_data <- prepareRawILOData(ilo_working_hours_data = ilo_working_hours_data,
                              ilo_employment_data = ilo_employment_data)
working_hours_sector_data <- hmw_data |>
  add_hmw_region_codes() |>
  fill_ilo_data() |>
  calc_total_hours_worked() |>
  get_broad.sector_data()