ILO data item code names

A string list containing a selection of ILO data code names. See Rilostat::get_ilostat_toc() to view the associated items, codes, and other metadata..

Usage

ilo_codes

Format

A string list with 2 entries.

working_hours_code

The string "HOW_TEMP_SEX_ECO_NB_A", representing Mean weekly hours actually worked per employed person by sex and economic activity.

employment_code

The string "EMP_TEMP_SEX_ECO_NB_A", representing Employment by sex and economic activity (thousands).

Examples

ilo_codes
#> $working_hours_code
#> [1] "HOW_TEMP_SEX_ECO_NB_A"
#> 
#> $employment_code
#> [1] "EMP_TEMP_SEX_ECO_NB_A"
#>