Economic sectors for muscle work

A list containing names of economic sectors for muscle work.

mw_sectors

Format

A string list with 13 entries.

broad_sector_col

The column name "broad_sector", which contains the "Broad sector" sector names used by the ILO.

hmw_analysis_sector_col

The column name "Sector.hmw", which contains the sector names used in the hmw analysis data excel file bundled with this package.

agriculture_broad.sector

The string "Agriculture" representing the agriculture sector. One of the ILO's "Broad sectors".

industry_broad.sector

The string "Industry" representing the industry sector. One of the ILO's "Broad sectors".

services_broad.sector

The string "Services" representing the services sector. One of the ILO's "Broad sectors".

not_classified_broad.sector

The string "Not classfied" representing unclassified labor data. One of the ILO's "Broad sectors".

non_agriculture_broad.sector

The string "Non-agriculture" representing all sectors outside of agriculture. One of the ILO's "Broad sectors".

total_sector

The string "Total" representing the entire economy. One of the ILO's "Broad sectors".

resources_sectors

The string "Resources" representing resouce sectors.

Examples

mw_sectors
#> $broad_sector_col
#> [1] "broad_sector"
#> 
#> $hmw_analysis_sector_col
#> [1] "Sector.hmw"
#> 
#> $agriculture_broad.sector
#> [1] "Agriculture"
#> 
#> $industry_broad.sector
#> [1] "Industry"
#> 
#> $services_broad.sector
#> [1] "Services"
#> 
#> $transport_sector
#> [1] "Transport"
#> 
#> $not_classified_broad.sector
#> [1] "Not classified"
#> 
#> $non_agriculture_broad.sector
#> [1] "Non-agriculture"
#> 
#> $total_sector
#> [1] "Total"
#> 
#> $resources_sector
#> [1] "Resources"
#> 
#> $farms
#> [1] "Farms"
#> 
#> $food_production
#> [1] "Food production"
#> 
#> $feed_production
#> [1] "Feed production"
#>