Create a phi vector for muscle work energy conversion chains

To move from an energy to an exergy description of an energy conversion chain, a vector of phi (exergy-to-energy ratio) values is needed. Entries in the vector are for each type of muscle work energy carrier.

Usage

phi_vec_mw(
  .phi_table = MWTools::phi_constants_mw,
  matrix_class = c("matrix", "Matrix"),
  mw_energy_carriers = MWTools::mw_products,
  product = MWTools::mw_cols$product,
  product_type = MWTools::row_col_types$product,
  phi = "phi"
)

Arguments

.phi_table

A table of phi (exergy-to-energy ratio) values.

matrix_class

The type of matrix to be created, one of "matrix" or "Matrix". Default is "matrix".

mw_energy_carriers

A vector of string energy carriers relevant to muscle work calculations. Default is MWTools::mw_products.

product

See MWTools::mw_cols.

product_type

See MWTools::row_col_types.

phi

The name of the phi column in the outgoing vector. Default is "phi".

Value

A column vector of phi values.

Examples

phi_vec_mw()
#>         phi
#> Food      1
#> Feed      1
#> Biomass   1
#> HuMech    1
#> AnMech    1
#> AnP       1
#> attr(,"rowtype")
#> [1] "Product"
#> attr(,"coltype")
#> [1] "phi"
phi_vec_mw(mw_energy_carriers = c("Food", "Feed"))
#>      phi
#> Food   1
#> Feed   1
#> attr(,"rowtype")
#> [1] "Product"
#> attr(,"coltype")
#> [1] "phi"