Row and column types for matrices made from muscle work data frames

A string list containing package constants used in MWTools package functions. This list is borrowed directly from the IEATools package.

Usage

row_col_types

Format

A string list with 6 entries.

See IEATools::row_col_types.

Examples

row_col_types
#> $industry
#> [1] "Industry"
#> 
#> $resource
#> [1] "Industry"
#> 
#> $sector
#> [1] "Industry"
#> 
#> $product
#> [1] "Product"
#> 
#> $unit
#> [1] "Unit"
#> 
#> $other
#> [1] "Other"
#>