Human muscle work (hmw) analysis data constants

A string list containing a selection of hmw analysis constants.

Usage

hmw_analysis_constants

Format

A string list with 19 entries.

total_wk_hrs_ilo_col

The data column "Total.hours [hours/year]" containing the number of hours worked yearly by all humans for a given country, sex, sector, and year.

col_1960

The column for 1960. Default is "1960".

col_2020

The column for 2020. Default is "2020".

labor_type_col

The data column "Labor.Type" containing information on the human labor type (Primary, Secondary, or Sedentary) of human workers.

labor_split_col

The data column "Labor.Type.Split" containing information on the proportion of human labor type by sector.

hmw_power_sheet

The excel sheet name "hmw_power", which contains information on the power output of humans performing muscle work.

hmw_food_sheet

The excel sheet name "hmw_food", which contains information on the food consumption of humans.

hmw_labor_map_sheet

The excel sheet name "hmw_sector_labor_map", which contains data on the split of labor type by ILO broad sector.

hmw_sector_map_sheet

The excel sheet name "hmw_sector_map", which contains information on the mapping the sector names used in the hmw analysis data excel file bundled with this package with the sector names in the ILO data.

hmw_plate_waste_sheet

The excel sheet name "hmw_plate_waste", which contains information on proportion of food waste by region.

hmw_harvest_waste_sheet

The excel sheet name "hmw_harvest_waste", which contains information on the proportion of phytomass lost at the harvest stage by region.

food_consumption_col

The data column "Food consumption [kcal/day per person]", which contains information on the daily food consumption of humans in kilocalories (kcal).

energy_pppa_col

The data column "Energy consumption [MJ/year per person]", which contains information on the yearly food consumption per working person in megajoules (MJ).

final_energy_col

The data column "Final energy [MJ/year]", which contains data on the yearly food consumption of all working persons in megajoules (MJ).

primary_energy_col

The data column "Primary energy [MJ/year]", which contains data on the yearly quantity of biomass required to meet the food requirements of all working persons in megajoules (MJ).

useful_energy_hmw_col

The data column "Useful energy [MJ/year]", which contains data on the yearly useful work performed by all working persons in megajoules (MJ).

power_col

The data column "Power [W]", which contains data on the power output of humans performing muscle work.

plate_waste_col

The data column "Plate waste [-]", which contains data on the proportion of food waste by region.

hmw_harvest_waste_col

The data column "arvest waste [-]", which contains data on the proportion of phytomass lost at the harvest stage by region.

Examples

hmw_analysis_constants
#> $total_wk_hrs_ilo_col
#> [1] "Total.hours [hours/year]"
#> 
#> $col_1960
#> [1] "1960"
#> 
#> $col_2020
#> [1] "2020"
#> 
#> $labor_type_col
#> [1] "Labor.Type"
#> 
#> $labor_split_col
#> [1] "Labor.Type.Split"
#> 
#> $hmw_power_sheet
#> [1] "hmw_power"
#> 
#> $hmw_food_sheet
#> [1] "hmw_food"
#> 
#> $hmw_sector_map_sheet
#> [1] "hmw_sector_map"
#> 
#> $hmw_labor_map_sheet
#> [1] "hmw_sector_labor_map"
#> 
#> $hmw_plate_waste_sheet
#> [1] "hmw_plate_waste"
#> 
#> $hmw_harvest_waste_sheet
#> [1] "hmw_harvest_waste"
#> 
#> $food_consumption_col
#> [1] "Food consumption [kcal/day per person]"
#> 
#> $energy_pppa_col
#> [1] "Energy consumption [MJ/year per person]"
#> 
#> $final_energy_col
#> [1] "Final energy [MJ/year]"
#> 
#> $primary_energy_col
#> [1] "Primary energy [MJ/year]"
#> 
#> $useful_energy_hmw_col
#> [1] "Useful energy [MJ/year]"
#> 
#> $power_col
#> [1] "Power [W]"
#> 
#> $plate_waste_col
#> [1] "Plate waste [-]"
#> 
#> $hmw_harvest_waste_col
#> [1] "Harvest waste [-]"
#>