Add energy type and method columns to tidy muscle work data frames

To specify muscle work data in preparation for conversion to PSUT matrices, the first step is specifying that these data are energy (not exergy) data and indicating the method for estimating the primary energy ("Biomass [from Resources]") associated with the initial stage of final energy ("Biomass"). We assume the Physical Content Method (PCM) in which the primary energy ("Biomass [from Resources]") is equal in magnitude to the final energy ("Biomass").

specify_energy_type_method(
  .hmw_df,
  .amw_df,
  energy_type = MWTools::mw_cols$energy_type,
  method = MWTools::mw_cols$method,
  e_type = MWTools::energy_types$e,
  pcm = MWTools::methods$pcm
)

Arguments

.hmw_df

A data frame produced by calc_hmw_pfu().

.amw_df

A data frame produced by calc_amw_pfu().

energy_type, method, e_type

See MWTools::mw_cols.

pcm

See MWTools::methods.

Value

A data frame in which energy_type and method columns are included.

Details

This function binds .hmw_df and .amw_df by rows.

Examples

hmw_df <- hmw_test_data_path() %>%
  read.csv() %>%
  calc_hmw_pfu()
amw_df <- amw_test_data_path() %>%
  read.csv() %>%
  calc_amw_pfu()
specify_energy_type_method(hmw_df, amw_df)
#> # A tibble: 2,988 × 9
#>    Country  Year Species       Stage   Sector   Unit    E.dot Energy.type Method
#>    <chr>   <dbl> <chr>         <chr>   <chr>    <chr>   <dbl> <chr>       <chr> 
#>  1 GBR      1969 Human females Final   Agricul… EJ    9.92e-4 E           PCM   
#>  2 GBR      1969 Human females Primary Agricul… EJ    1.60e-3 E           PCM   
#>  3 GBR      1969 Human females Useful  Agricul… EJ    5.91e-5 E           PCM   
#>  4 GBR      1970 Human females Final   Agricul… EJ    9.70e-4 E           PCM   
#>  5 GBR      1970 Human females Primary Agricul… EJ    1.56e-3 E           PCM   
#>  6 GBR      1970 Human females Useful  Agricul… EJ    5.72e-5 E           PCM   
#>  7 GBR      1971 Human females Final   Agricul… EJ    9.32e-4 E           PCM   
#>  8 GBR      1971 Human females Primary Agricul… EJ    1.50e-3 E           PCM   
#>  9 GBR      1971 Human females Useful  Agricul… EJ    5.49e-5 E           PCM   
#> 10 GBR      1972 Human females Final   Agricul… EJ    9.07e-4 E           PCM   
#> # … with 2,978 more rows